Avoid pointers truncation on x64. Patch by JanWillem32. This commit fixes ticket #2286.
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@4701 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/DSUtil/DSMPropertyBag.h
@@ -99,7 +99,7 @@ public:
 
 	// global access to all resources
 	static CCritSec m_csResources;
-	static CAtlMap<DWORD, CDSMResource*> m_resources;
+	static CAtlMap<uintptr_t, CDSMResource*> m_resources;
 };
 
 class IDSMResourceBagImpl : public IDSMResourceBag
